Feature/refactoring

+++ src/main/java/com/saxatus/aiml/tags/LearnTag.java
@@ -3,30 +3,26 @@ package com.saxatus.aiml.tags;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.PrintStream;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.Result;
-import javax.xml.transform.Source;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.TransformerException;
 
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
+import org.xml.sax.SAXException;
 
 import com.saxatus.aiml.factory.TagFactory;
 import com.saxatus.aiml.parsing.AIMLParseNode;
+import com.saxatus.aiml.utils.XMLUtils;
 
 public class LearnTag extends AbstractBotTag
 {
     // TODO: rework
-
     private LearnTag(Node node, TagFactory factory)
     {
         super(node, factory);
@@ -88,11 +84,11 @@ public class LearnTag extends AbstractBotTag
                             .getName())
                             .log(Level.SEVERE, "Exception in LearnTag.handle: ", e);
         }
-        // bot.setReinitFlag(true);
         return "";
     }
 
-    private void updateAIML(String pattern, String template) throws Exception
+    private void updateAIML(String pattern, String template)
+                    throws IOException, ParserConfigurationException, SAXException, TransformerException
     {
         File learnFile = getAIMLHandler().getLearnFile();
         if (!learnFile.getParentFile()
@@ -101,9 +97,8 @@ public class LearnTag extends AbstractBotTag
             learnFile.getParentFile()
                             .mkdirs();
         }
-        if (!learnFile.exists())
+        if (!learnFile.exists() && learnFile.createNewFile())
         {
-            learnFile.createNewFile();
             try (PrintStream out = new PrintStream(new FileOutputStream(learnFile)))
             {
                 out.print("<aiml></aiml>");
@@ -112,11 +107,9 @@ public class LearnTag extends AbstractBotTag
             {
                 e.printStackTrace();
             }
+
         }
-        DocumentBuilderFactory dbFactory = DocumentBuilderFactory.newInstance();
-        DocumentBuilder dBuilder;
-        dBuilder = dbFactory.newDocumentBuilder();
-        Document doc = dBuilder.parse(learnFile);
+        Document doc = XMLUtils.parseFileToXMLDocument(learnFile);
         Node node = doc.createElement("category");
         node.appendChild(doc.createElement("pattern"))
                         .appendChild(doc.createTextNode(pattern));
@@ -126,11 +119,7 @@ public class LearnTag extends AbstractBotTag
         doc.getElementsByTagName("aiml")
                         .item(0)
                         .appendChild(node);
-        Transformer transformer = TransformerFactory.newInstance()
-                        .newTransformer();
-        Result output = new StreamResult(learnFile);
-        Source input = new DOMSource(doc);
-        transformer.transform(input, output);
+        XMLUtils.writeXMLDocumentToFile(doc, learnFile);
     }
 
 }
